Move endpoints schema and hash functions to schema_endpoints.go
